refactor: consolidate unimplemented service stubs into service.go
