+++ category.go
@@ -1,10 +1,1 @@
 package backlog
-
-import "github.com/nattokin/go-backlog/internal/core"
-
-// CategoryService handles communication with the category-related methods of the Backlog API.
-//
-//nolint:unused // API not implemented yet
-type CategoryService struct {
-	method *core.Method
-}
