+++ customfield.go
@@ -1,10 +1,1 @@
 package backlog
-
-import "github.com/nattokin/go-backlog/internal/core"
-
-// CustomFieldService handles communication with the custom field-related methods of the Backlog API.
-//
-//nolint:unused // API not implemented yet
-type CustomFieldService struct {
-	method *core.Method
-}
