+++ priority.go
@@ -1,10 +1,1 @@
 package backlog
-
-import "github.com/nattokin/go-backlog/internal/core"
-
-// PriorityService handles communication with the priority-related methods of the Backlog API.
-//
-//nolint:unused // API not implemented yet
-type PriorityService struct {
-	method *core.Method
-}
