+++ repository.go
@@ -1,10 +1,1 @@
 package backlog
-
-import "github.com/nattokin/go-backlog/internal/core"
-
-// RepositoryService handles communication with the repository-related methods of the Backlog API.
-//
-//nolint:unused // API not implemented yet
-type RepositoryService struct {
-	method *core.Method
-}
