+++ resolution.go
@@ -1,10 +1,1 @@
 package backlog
-
-import "github.com/nattokin/go-backlog/internal/core"
-
-// ResolutionService handles communication with the resolution-related methods of the Backlog API.
-//
-//nolint:unused // API not implemented yet
-type ResolutionService struct {
-	method *core.Method
-}
