+++ status.go
@@ -1,10 +1,1 @@
 package backlog
-
-import "github.com/nattokin/go-backlog/internal/core"
-
-// StatusService handles communication with the status-related methods of the Backlog API.
-//
-//nolint:unused // API not implemented yet
-type StatusService struct {
-	method *core.Method
-}
